Big file import, clear modal #40 #41 #42 #43
Now the user can import a big file without the browser crashing.
Import done in batches, and the user can see the progress.
Added a modal to ask the user if they want to clear the database before importing a new file.
Added the jar file to the assets folder which is used in the advanced queries, and needed as a plugin for the neo4j database.
Added a new setting to user to decide the size of batches when importing a file.
Getting the paths and walks from the database needs to be updated.